The use of the 'describe' verb is not valid.
There is no validation of verb names in Kubernetes so this won't cause
an error but it also won't do anything. The 'get' verb is enough for
getting the jobs and their status.

It will generate an error if the account creating the role has been
restricted in what it can create.